pyproject.toml
[project]
name = "proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.73.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.9",
    "openai>=2.32.0",
    "pandas>=3.0.2",
    "pydantic-settings>=2.14.0",
    "pymupdf>=1.27.2.3",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "wordcloud>=1.9.6",
]
